Introduce "botApi" -module which handles all actions towards Telegram API
Otherwise the upcoming userController would have created circular dependency loop
