+++ botApi.js
@@ -1,0 +1,67 @@
+var cfg             = require('./config');
+var request         = require('request');
+
+var botApi = {};
+
+// ## Public functions
+//
+
+botApi.sendMessage = function(chatId, text) {
+    request.post(cfg.tgApiUrl + '/sendMessage', { form: {
+        chat_id: chatId,
+        text: text
+    }});
+};
+
+botApi.sendPhoto = function (chatId, photo, options) {
+    var opts = {
+        qs: options || {}
+    };
+    opts.qs.chat_id = chatId;
+    var content = _formatSendData('photo', photo);
+
+    opts.formData = content.formData;
+    opts.qs.photo = content.file;
+    request.post(cfg.tgApiUrl + '/sendPhoto', opts);
+};
+
+
+
+// ## Internal functions
+//
+
+var _formatSendData = function (type, data) {
+    var formData = {};
+    var fileName;
+    var fileId = data;
+
+    if (data instanceof stream.Stream) {
+        fileName = path.basename(data.path);
+
+        formData[type] = {
+            value: data,
+            options: {
+                filename: fileName,
+                contentType: mime.lookup(fileName)
+            }
+        };
+    }
+    else if (fs.existsSync(data)) {
+        fileName = path.basename(data);
+
+        formData[type] = {
+            value: fs.createReadStream(data),
+            options: {
+                filename: fileName,
+                contentType: mime.lookup(fileName)
+            }
+        };
+    }
+
+    return {
+        formData: formData,
+        file: fileId
+    };
+};
+
+module.exports = botApi;
